Fix cycle in webview contrib imports  (#163712)
Fix cycle in webview contrib

The top level `webview.ts` should not import from `webviewElement`
